test(zod-utils): add coverage for friendly error map
